Move create account into AccountList class, modify AccountList constructor to create account if savefile does not exist automatically
