+++ src/main/java/seedu/bankwithus/exceptions/AccountNotFoundException.java
@@ -1,0 +1,5 @@
+package seedu.bankwithus.exceptions;
+
+public class AccountNotFoundException extends Exception {
+    
+}
